Renamed folders and added information
Renamed the 'example1-1'folder to 'Chapter 1' so it's consistent with the book. Added informaiton about each file in Chapter1 files (which file is which on the book). Updated the README file with the link to the original book code. Updated main.cc (example1.1 on the book) with some comments for better understanding.